pyproject.toml
[tool.poetry]
name = "english-class-notify"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
logzero = "^1.5.0"
black = "^19.10b0"
pytest = "^6.0.1"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
